The Nobel Prize is perhaps the world's most well known scientific award. Every year it is given to scientists and scholars in chemistry, literature, physics, medicine, economics, and peace. The first Nobel Prize was handed out in 1901, and at that time the prize was Eurocentric and male-focused, but nowadays it's not biased in any way. Surely, right?
Project
 A Visual History of Nobel Prize Winners

The Nobel Prize is perhaps the world's most well known scientific award. Every year it is given to scientists and scholars in chemistry, literature, physics, medicine, economics, and peace. The first Nobel Prize was handed out in 1901, and at that time the prize was Eurocentric and male-focused, but nowadays it's not biased in any way. Surely, right?

Well, let's find out! In this project, you get to explore patterns and trends in over 100 years worth of Nobel Prize winners. What characteristics do the prize winners have? Which country gets it most often? And has anybody gotten it twice? It's up to you to figure this out.

Before starting this project, we recommend that you have completed:

pandas Foundations
Data Visualization with Seaborn
The dataset used in this project is from The Nobel Foundation on Kaggle.

Project Tasks

The most Nobel of Prizes
So, who gets the Nobel Prize?
USA dominance
USA dominance, visualized
What is the gender of a typical Nobel Prize winner?
The first woman to win the Nobel Prize
Repeat laureates
How old are you when you get the prize?
Age differences between prize categories
Oldest and youngest winners
You get a prize!